pkg/endpoint: move ModifyIdentityLabels to identity.go
Signed-off by: Ian Vernon <ian@cilium.io>
